tts-test/pyproject.toml
[project]
name = "tts-test"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "azure-cognitiveservices-speech>=1.44.0",
    "boto3>=1.38.25",
    "google-cloud-texttospeech>=2.27.0",
    "python-dotenv>=1.1.0",
]
